Move Favorite Ski Resort
Move implementation of favorite ski resort to its own dedicated menu inside of Ski Profile
