+++ src/main/java/de/dennisguse/opentracks/settings/skiProfile/SkiProfileSharpeningFragment.java
@@ -1,4 +1,4 @@
-package de.dennisguse.opentracks.settings;
+package de.dennisguse.opentracks.settings.skiProfile;
 
 import android.os.Bundle;
 import android.text.InputFilter;
@@ -16,12 +16,12 @@ import android.content.SharedPreferences;
 import de.dennisguse.opentracks.R;
 
 
-public class SkiProfileWaxingFragment extends PreferenceFragmentCompat {
+public class SkiProfileSharpeningFragment extends PreferenceFragmentCompat {
     @Override
     public void onCreatePreferences(@Nullable Bundle savedInstanceState, @Nullable String rootKey) {
-        setPreferencesFromResource(R.xml.fragment_ski_profile_waxing_info, rootKey);
+        setPreferencesFromResource(R.xml.fragment_ski_profile_sharpening_info, rootKey);
 
-        EditTextPreference waxingIntervalPreference = findPreference("waxing_interval");
+        EditTextPreference waxingIntervalPreference = findPreference("sharpening_interval");
         if (waxingIntervalPreference != null) {
             waxingIntervalPreference.setOnBindEditTextListener(new EditTextPreference.OnBindEditTextListener() {
                 @Override
@@ -44,48 +44,25 @@ public class SkiProfileWaxingFragment extends PreferenceFragmentCompat {
             waxingIntervalPreference.setSummary(defaultValue + " km");
         }
 
-        ListPreference waxingTypePreference = findPreference("type_of_wax");
-        if (waxingTypePreference != null) {
-            waxingTypePreference.setOnPreferenceChangeListener(new Preference.OnPreferenceChangeListener() {
-                @Override
-                public boolean onPreferenceChange(Preference preference, Object newValue) {
-                    String newValueString = (String) newValue;
-                    int index = waxingTypePreference.findIndexOfValue(newValueString);
-                    if (index >= 0) {
-                        String entry = waxingTypePreference.getEntries()[index].toString();
-                        preference.setSummary(entry);
-                    }
-                    return true;
-                }
-            });
 
-            // Update summary initially with the default value
-            String defaultValue = waxingTypePreference.getValue();
-            int index = waxingTypePreference.findIndexOfValue(defaultValue);
-            if (index >= 0) {
-                String entry = waxingTypePreference.getEntries()[index].toString();
-                waxingTypePreference.setSummary(entry);
-            }
-        }
 
-        Preference lastWaxingDatePreference = findPreference("last_waxing_date");
+        Preference lastWaxingDatePreference = findPreference("last_sharpening_date");
 
         // Retrieve the last selected date from SharedPreferences
         SharedPreferences sharedPreferences = getPreferenceManager().getSharedPreferences();
-        String lastSelectedDate = sharedPreferences.getString("last_waxing_date", "");
+        String lastSelectedDate = sharedPreferences.getString("last_sharpening_date", "");
 
         lastWaxingDatePreference.setSummary(lastSelectedDate);
-            lastWaxingDatePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
-                @Override
-                public boolean onPreferenceClick(Preference preference) {
-                    showDatePickerDialog();
-                    return true;
-                }
-            });
+        lastWaxingDatePreference.setOnPreferenceClickListener(new Preference.OnPreferenceClickListener() {
+            @Override
+            public boolean onPreferenceClick(Preference preference) {
+                showDatePickerDialog();
+                return true;
+            }
+        });
 
-        
+
     }
-
 
     // Validation for Waxing interval field
     private class DecimalInputFilter implements InputFilter {
@@ -122,21 +99,21 @@ public class SkiProfileWaxingFragment extends PreferenceFragmentCompat {
                     @Override
                     public void onDateSet(DatePicker view, int year, int month, int dayOfMonth) {
                         String selectedDate = dayOfMonth + "/" + (month + 1) + "/" + year;
-                        Preference lastWaxingDatePreference = findPreference("last_waxing_date");
+                        Preference lastWaxingDatePreference = findPreference("last_sharpening_date");
                         lastWaxingDatePreference.setSummary(selectedDate);
 
                         // Handle the selected date here, for example, save it to SharedPreferences
                         SharedPreferences.Editor editor = getPreferenceManager().getSharedPreferences().edit();
-                        editor.putString("last_waxing_date", selectedDate);
+                        editor.putString("last_sharpening_date", selectedDate);
                         editor.apply();
                     }
                 },
                 year, month, dayOfMonth
         );
-        
+
         // Set maximum date limit to prevent selection of future dates
         datePickerDialog.getDatePicker().setMaxDate(System.currentTimeMillis());
-        
+
         datePickerDialog.show();
     }
