+++ src/main/res/xml/fragment_ski_profile_sharpening_info.xml
@@ -3,30 +3,6 @@
     <PreferenceCategory
         android:title="Last Sharpening Date">
 
-<!--        <ListPreference-->
-<!--            android:key="year"-->
-<!--            android:title="Year"-->
-<!--            android:summary="Choose the year of the last sharpening date"-->
-<!--            android:entries="@array/time_year"-->
-<!--            android:entryValues="@array/time_year"-->
-<!--            android:defaultValue="yearly" />-->
-
-<!--        <ListPreference-->
-<!--            android:key="month"-->
-<!--            android:title="Monty"-->
-<!--            android:summary="Choose the month of the last sharpening date"-->
-<!--            android:entries="@array/time_month"-->
-<!--            android:entryValues="@array/time_month"-->
-<!--            android:defaultValue="monthly" />-->
-
-<!--        <ListPreference-->
-<!--            android:key="day"-->
-<!--            android:title="Day"-->
-<!--            android:summary="Choose the day of the last sharpening date"-->
-<!--            android:entries="@array/time_day"-->
-<!--            android:entryValues="@array/time_day"-->
-<!--            android:defaultValue="daily" />-->
-
         <Preference
             android:key="last_sharpening_date"
             android:title="Select Date"
@@ -45,7 +21,6 @@
             android:dialogTitle="Enter sharpening Interval"
             android:inputType="numberDecimal"
             android:defaultValue="30" />
-
 
         <ListPreference
             android:key="base_angle"
